Add _atoi function

+++ pointers_arrays_strings/100-atoi.c
@@ -32,5 +32,5 @@ int _atoi(char *s)
 	}
 
 	/* answer calculation */
-	return (sign * number);
+	return ((int)(sign * number));
 }
